src/main/java/org/radio/Radio.java
package org.radio;

public class Radio {
    private int currentVolume;
    private int currentStation;

    public Radio() {
        this.currentStation = 0;
        this.currentVolume = 0;
    }

    public Radio(int initialStation, int initialVolume) {
        this.currentStation = initialStation;
        this.currentVolume = initialVolume;
    }

    public void setStation(int station) {
        if (station >= 0 && station <= 9) {
            currentStation = station;
        }
    }

    public int getCurrentStation() {
        return currentStation;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }

    public void next() {
        if (currentStation == 9) {
            currentStation = 0;
        } else {
            currentStation++;
        }
    }

    public void prev() {
        if (currentStation == 0) {
            currentStation = 9;
        } else {
            currentStation--;
        }
    }

    public void increaseVolume() {
        if (currentVolume < 10) {
            currentVolume++;
        }
    }

    public void decreaseVolume() {
        if (currentVolume > 0) {
            currentVolume--;
        }
    }
}
